Melhorias

+++ src/main/java/br/com/fiap/geradorCupons/persistense/model/Item.java
@@ -35,7 +35,7 @@ public class Item implements Serializable {
 		this.id = id;
 	}
 
-	public double getQuantidade() {
+	public int getQuantidade() {
 		return quantidade;
 	}
